# server/public/swagger.json
{
	"components": {
...
		"schemas": {
			"IUserName": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"fname": {
						"type": "string"
					},
					"lname": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"fname",
					"lname"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Status": {
				"enum": [
					"ordering",
					"ordered",
					"delivery",
					"canceled",
					"completed"
				],
				"type": "string"
			},
			"PaymentMethod": {
				"enum": [
					"cod",
					"banktransfer",
					"e-wallet",
					"gateway"
				],
				"type": "string"
			},
			"IOrderSearchProps": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"paymentCheck": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"shipping": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					},
					"orderAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/IUserName"
					}
				},
				"required": [
					"id",
					"status",
					"paymentMethod",
					"paymentCheck",
					"address",
					"total",
					"orderAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrders": {
				"properties": {
					"orders": {
						"items": {
							"$ref": "#/components/schemas/IOrderSearchProps"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"orders",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Period": {
				"enum": [
					"today",
					"week",
					"month",
					"quarter",
					"year"
				],
				"type": "string"
			},
			"OrderField": {
				"enum": [
					"orderAt",
					"total"
				],
				"type": "string"
			},
			"Change": {
				"enum": [
					"DESC",
					"ASC"
				],
				"type": "string"
			},
			"UserRole": {
				"enum": [
					"all",
					"admin",
					"editor",
					"customer"
				],
				"type": "string"
			},
			"IUserAuth": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"fname": {
						"type": "string"
					},
					"lname": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"id",
					"fname",
					"lname",
					"phone",
					"address",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IItemDetails": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"productVariantId": {
						"type": "number",
						"format": "double"
					},
					"product": {
						"properties": {
							"size": {
								"type": "string"
							},
							"color": {
								"type": "string"
							},
							"brand": {
								"type": "string"
							},
							"imgPaths": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"size",
							"color",
							"brand",
							"imgPaths",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"orderId",
					"quantity",
					"unitPrice",
					"productVariantId",
					"product"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrderInfo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"paymentCheck": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"shipping": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					},
					"orderAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/IUserAuth"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/IItemDetails"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"status",
					"paymentMethod",
					"paymentCheck",
					"address",
					"total",
					"orderAt",
					"user",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateOrderDetailModel": {
				"properties": {
					"orderId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"productVariantId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"quantity": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"unitPrice": {
						"type": "number",
						"format": "double",
						"minimum": 0
					}
				},
				"required": [
					"productVariantId",
					"quantity",
					"unitPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateOrderCreateModel": {
				"properties": {
					"details": {
						"items": {
							"$ref": "#/components/schemas/ValidateOrderDetailModel"
						},
						"type": "array"
					},
					"userId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					}
				},
				"required": [
					"details",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrderBasicProps": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"paymentCheck": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"shipping": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					},
					"orderAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status",
					"paymentMethod",
					"paymentCheck",
					"address",
					"total",
					"orderAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateOrderUpdateModel": {
				"properties": {
					"details": {
						"items": {
							"$ref": "#/components/schemas/ValidateOrderDetailModel"
						},
						"type": "array"
					}
				},
				"required": [
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrderUpdateProps": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"paymentCheck": {
						"type": "boolean"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"address": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IPlaceOrder": {
				"properties": {
					"address": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"shipping": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"address",
					"paymentMethod"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrderDetail": {
				"properties": {
					"orderId": {
						"type": "number",
						"format": "double"
					},
					"productVariantId": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"productVariantId",
					"quantity",
					"unitPrice",
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateUpdateQuantityModel": {
				"properties": {
					"quantity": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					}
				},
				"required": [
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IOrderReport": {
				"properties": {
					"completedOrders": {
						"type": "number",
						"format": "double"
					},
					"canceledOrders": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"completedOrders",
					"canceledOrders"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Color": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"code",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateColorModel": {
				"properties": {
					"name": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"code": {
						"type": "string",
						"maxLength": 7,
						"pattern": "^[#]\\w{6}"
					}
				},
				"required": [
					"name",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "cookie",
				"name": "token"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "belle-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LinhTran"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/orders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrders"
								}
							}
						}
					}
				},
				"description": "For Admin permission\r\nGet all Orders, be able to sort by price, time, user and to filter by price, time, user...",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "time",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Period"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Status"
						}
					},
					{
						"in": "query",
						"name": "paymentCheck",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderField"
						}
					},
					{
						"in": "query",
						"name": "change",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Change"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderInfo"
								}
							}
						}
					}
				},
				"description": "Create new order when user add items to order that is not already existed.",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateOrderCreateModel"
							}
						}
					}
				}
			}
		},
		"/orders/{userId}/all": {
			"get": {
				"operationId": "GetOrdersOfUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrderBasicProps"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all orders of user by userId",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
		"/orders/{userId}/current": {
			"get": {
				"operationId": "GetCurrentOrderOfUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IOrderInfo"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Current order is incomplete order (status: ORDERING)\r\nGet current order of user",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
		"/orders/{orderId}": {
			"get": {
				"operationId": "GetOrderById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderInfo"
								}
							}
						}
					}
				},
				"description": "Get order details",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteOrderById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete order of user",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
		"/orders/{userId}/afterLogin": {
			"patch": {
				"operationId": "UpdateOrderItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderInfo"
								}
							}
						}
					}
				},
				"description": "Update order (add items to existing cart or create new cart when users login if there are items in their carts)",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateOrderUpdateModel"
							}
						}
					}
				}
			}
		},
		"/orders/{orderId}/updateStatus": {
			"patch": {
				"operationId": "UpdateOrderStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderInfo"
								}
							}
						}
					}
				},
				"description": "Update order status",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrderUpdateProps"
							}
						}
					}
				}
			}
		},
		"/orders/{orderId}/place": {
			"patch": {
				"operationId": "PlaceOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderInfo"
								}
							}
						}
					}
				},
				"description": "Submit current order",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPlaceOrder"
							}
						}
					}
				}
			}
		},
		"/orders/{orderId}/items": {
			"post": {
				"operationId": "AddItemToOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderInfo"
								}
							}
						}
					}
				},
				"description": "Add (create) item to order",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateOrderDetailModel"
							}
						}
					}
				}
			}
		},
		"/orders/items/{itemId}": {
			"patch": {
				"operationId": "UpdateItemQuantity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderDetail"
								}
							}
						}
					}
				},
				"description": "Update order item (quantity)",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateUpdateQuantityModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteItem",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete order items",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"customer",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
		"/report/orders": {
			"get": {
				"operationId": "GetOrderReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderReport"
								}
							}
						}
					}
				},
				"description": "Get the number of completed orders and canceled order",
				"tags": [
					"Report Controller"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "time",
						"required": true,
						"schema": {
							"type": "string",
							"pattern": "(^[\\d]{4}$)|(^[\\d]{4}-([0][1-9]|[1][0-2])|(^week$))"
						}
					}
				]
			}
		},
...
		"/colors": {
			"get": {
				"operationId": "GetColors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Color"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all colors",
				"tags": [
					"Product Color"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateColor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Color"
								}
							}
						}
					}
				},
				"description": "Create new color",
				"tags": [
					"Product Color"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateColorModel"
							}
						}
					}
				}
			}
		},
		"/colors/{id}": {
			"patch": {
				"operationId": "UpdateColorById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Color"
								}
							}
						}
					}
				},
				"description": "Update color info",
				"tags": [
					"Product Color"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateColorModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteColorById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete color",
				"tags": [
					"Product Color"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}